Finshed Adding MissingOrder functions finally , complated insertOrder , GetProductInfo functions in manageDB , some change to AddRemoveEdit window design for easier management , fixed a bug where app would crush if the user left any textbox empty in add or remove anything

+++ DashBoardDB/AddRemoveEdIt.xaml.cs
@@ -23,6 +23,7 @@ namespace DashBoardDB
         List<String> CompaniesList = new List<String>();
         List<String> MissingOrderNames = new List<String>();
         List<int> MissingOrderQuanties = new List<int>();
+        List<Double> MissingOrderPfPriceForEachProduct = new List<Double>();
         String MissingOrderProductName = "";
         Double MissingOrderTotalPrice = 0;
         public AddRemoveEdIt()
@@ -94,6 +95,12 @@ namespace DashBoardDB
             CompanyBox.Visibility = hide;
             expdateblock.Visibility = hide;
             expdatepciker.Visibility = hide;
+            OrderProductNameBlock.Visibility = hide;
+            OrderProductNameBox.Visibility = hide;
+            OrderProductQuantityBlock.Visibility = hide;
+            OrderProductQuantityBox.Visibility = hide;
+            ScrollBar.Visibility = hide;
+            AddProductToOrderBlock.Visibility = hide;
         }
         private void PrepareAddProduct()
         {
@@ -117,9 +124,11 @@ namespace DashBoardDB
                 productpfbricebox.Visibility = Show;
                 quantityblock.Visibility = Show;
                 quantitybox.Visibility = Show;
+                quantitybox.Text = "Quantity:";
                 companyblock.Visibility = Show;
                 CompanyBox.Visibility = Show;
                 expdateblock.Visibility = Show;
+                expdateblock.Text = "Exp date:";
                 expdatepciker.Visibility = Show;
                 expdatepciker.Visibility = Show;
                 PhoneNumber.Visibility = hide;
@@ -191,11 +200,14 @@ namespace DashBoardDB
                 ScrollBar.IsEnabled = true;
                 AddProductToOrderBlock.Visibility = Show;
                 ProductNameAndQuantityBlock.Visibility = Show;
-                quantityblock.Text = "Price";
+                quantityblock.Text = "Price:";
                 quantityblock.Visibility = Show;
                 quantitybox.Visibility = Show;
                 quantitybox.Text = "";
                 quantitybox.IsReadOnly = true;
+                expdateblock.Text = "Date:";
+                expdateblock.Visibility = Show;
+                expdatepciker.Visibility = Show;
                 Confirm.Visibility = Show;
             }
             else
@@ -214,6 +226,8 @@ namespace DashBoardDB
                 ScrollBar.Visibility = hide;
                 ScrollBar.IsEnabled = false;
                 quantitybox.IsReadOnly = false;
+                expdateblock.Visibility = hide;
+                expdatepciker.Visibility = hide;
             }
         }
         private void PrepareRemoveProdect()
@@ -224,6 +238,7 @@ namespace DashBoardDB
             {
                 hideboxes();
                 productpfpriceblock.Text = "        Name:";
+                productpfpriceblock.Margin = new Thickness(2, 4, 318, 18);
                 productpfpriceblock.Visibility = Show;
                 productpfbricebox.Visibility = Show;
                 Confirm.Visibility = Show;
@@ -239,6 +254,7 @@ namespace DashBoardDB
             {
                 hideboxes();
                 productpfpriceblock.Text = "        Name:";
+                productpfpriceblock.Margin = new Thickness(2, 4, 318, 18);
                 productpfpriceblock.Visibility = Show;
                 productpfbricebox.Visibility = Show;
                 Confirm.Visibility = Show;
@@ -376,6 +392,7 @@ namespace DashBoardDB
                 }
                 MissingOrderNames.Add(MissingOrderProductName);
                 MissingOrderQuanties.Add(MissingOrderProductQuantity);
+                MissingOrderPfPriceForEachProduct.Add(MissingOrderTotalPriceForLastProduct / MissingOrderProductQuantity);
                 ProductNameAndQuantityBlock.Text += Environment.NewLine + "      " + MissingOrderProductName + "                                                  " + MissingOrderProductQuantity.ToString();
                 OrderProductNameBox.Text = "";
                 OrderProductQuantityBox.Text = "";
@@ -389,6 +406,12 @@ namespace DashBoardDB
             {
                 if (EditType.SelectedIndex == 1)
                 {
+                    if (productnamebox.Text == "" || productpfbricebox.Text == "" || PhoneNumbox.Text == "" || TypeBox.SelectedItem.ToString() == "" ||
+                        CompanyBox.SelectedItem.ToString() == "" || quantitybox.Text == "" || productojbricebox.Text == "" || expdatepciker.SelectedDate.Value.ToString() == "")
+                    {
+                        MessageBox.Show("Dont leave any TextBox empty!");
+                        return;
+                    }
                     Checker = managaDB.insertProduct(productnamebox.Text, TypeBox.SelectedItem.ToString(), CompanyBox.SelectedItem.ToString(),
                         Convert.ToInt32(quantitybox.Text), Convert.ToDouble(productojbricebox.Text), Convert.ToDouble(productpfbricebox.Text), expdatepciker.SelectedDate.Value);
                     if (Checker == true)
@@ -396,35 +419,78 @@ namespace DashBoardDB
                     else
                         MessageBox.Show("There was an error adding the product to the DB");
                 }
-                if (EditType.SelectedIndex == 2)
+                else if (EditType.SelectedIndex == 2)
                 {
+                    if (productnamebox.Text == "" || productpfbricebox.Text == "" || PhoneNumbox.Text == "")
+                    {
+                        MessageBox.Show("Dont leave any TextBox empty!");
+                        return;
+                    }    
                     Checker = managaDB.InsertCompany(productnamebox.Text, productpfbricebox.Text, Convert.ToDouble(PhoneNumbox.Text));
                     if (Checker == true)
                         MessageBox.Show("Company has been added to the DB");
                     else
                         MessageBox.Show("There was an error adding the company to the DB (hint:maybe the name is wrong)");
                 }
+                else if (EditType.SelectedIndex == 3)
+                {
+                    if(quantitybox.Text == "")//if quantityBox is empty then they didnt any product to the list yet so check it alone is enough
+                    {
+                        MessageBox.Show("Add product to the list first!");
+                        return;
+                    }
+                    else if(expdatepciker.SelectedDate.Value == null)
+                    {
+                        MessageBox.Show("Pick a date First!");
+                        return;
+                    }
+                    else if(expdatepciker.SelectedDate.Value.Date > DateTime.Today)
+                    {
+                        MessageBox.Show("Date cant be after now!");//bad wording but idk how to say it!
+                        return;
+                    }
+                    Checker = managaDB.InsertOrder(MissingOrderNames, MissingOrderQuanties,MissingOrderPfPriceForEachProduct, MissingOrderTotalPrice, expdatepciker.SelectedDate.Value);
+                    if (Checker == true)
+                        MessageBox.Show("Order has been added to the DB");
+                    else
+                        MessageBox.Show("There was an error while adding the order to the DB");
+                }
             }
             else
             {
                 if (EditType.SelectedIndex == 1)
                 {
+                    if(productpfbricebox.Text == "")
+                    {
+                        MessageBox.Show("Dont leave the TextBox empty!");
+                        return;
+                    }
                     Checker = managaDB.DeleteRecord("products", "ProductName", productpfbricebox.Text);
                     if (Checker == true)
                         MessageBox.Show("Product has been removed from the DB");
                     else
                         MessageBox.Show("There was an error removeing the product from the DB");
                 }
-                if (EditType.SelectedIndex == 2)
+                else if (EditType.SelectedIndex == 2)
                 {
+                    if (productpfbricebox.Text == "")
+                    {
+                        MessageBox.Show("Dont leave the TextBox empty!");
+                        return;
+                    }
                     Checker = managaDB.DeleteRecord("companies", "CompanyName", productpfbricebox.Text);
                     if (Checker == true)
                         MessageBox.Show("Company has been removed from the DB");
                     else
                         MessageBox.Show("There was an error removeing the company from the DB (hint:maybe the name is wrong)");
                 }
-                if (EditType.SelectedIndex == 3)
+                else if(EditType.SelectedIndex == 3)
                 {
+                    if (productpfbricebox.Text == "")
+                    {
+                        MessageBox.Show("Dont leave the TextBox empty!");
+                        return;
+                    }
                     Checker = managaDB.DeleteRecord("orders", "idOrders", productpfbricebox.Text);
                     if (Checker == true)
                         MessageBox.Show("The order has been removed from the DB");
